+++ ContextualPopupDecorator/ContextualPopupDecorator.js
@@ -665,9 +665,13 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.positionContextualPopup();
 			const current = Spotlight.getCurrent();
 			this.updateLeaveFor(current);
-			this.setState({
-				activator: current
-			}, () => this.spotPopupContent());
+
+			setTimeout(() => {
+				this.setState({
+					activator: current
+				});
+				this.spotPopupContent();
+			});
 		};
 
 		handleClose = () => {